implemented triangle

+++ src/Ellipse.kt
@@ -3,7 +3,7 @@ open class Ellipse(_points: MutableList<Point>, _radii: MutableList<Double>): Sh
     var area = Math.PI * radii[0] * radii[1]
 
     private fun checkArea(): Boolean {
-        if (area.toInt() == 0){
+        if (area == 0.toDouble()){
             return false
         }
         return true
